Move crown_r0 and crown_zmax allometry functions into t_model_functions
